+++ pipecat-client-android/src/main/java/ai/pipecat/client/result/Callback.kt
@@ -1,0 +1,11 @@
+package ai.pipecat.client.result
+
+/**
+ * A callback function receiving a value.
+ */
+fun interface Callback<T> {
+    fun onComplete(value: T)
+}
+
+
+