+++ pipecat-client-android/src/main/java/ai/pipecat/client/types/MediaDeviceId.kt
@@ -1,0 +1,8 @@
+package ai.pipecat.client.types
+
+/**
+ * A unique identifier for a media device.
+ */
+data class MediaDeviceId(
+    val id: String
+)