+++ pipecat-client-android/src/main/java/ai/pipecat/client/types/MediaDeviceInfo.kt
@@ -1,0 +1,9 @@
+package ai.pipecat.client.types
+
+/**
+ * Information about a media device.
+ */
+data class MediaDeviceInfo(
+    val id: MediaDeviceId,
+    val name: String,
+)