+++ pipecat-client-android/src/main/java/ai/pipecat/client/types/OptionDescription.kt
@@ -1,0 +1,9 @@
+package ai.pipecat.client.types
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class OptionDescription(
+    val name: String,
+    val type: Type
+)