+++ pipecat-client-android/src/main/java/ai/pipecat/client/types/ParticipantId.kt
@@ -1,0 +1,8 @@
+package ai.pipecat.client.types
+
+/**
+ * A unique identifier for a session participant.
+ */
+data class ParticipantId(
+    val id: String
+)