+++ pipecat-client-android/src/main/java/ai/pipecat/client/types/ServiceConfig.kt
@@ -1,0 +1,9 @@
+package ai.pipecat.client.types
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ServiceConfig(
+    val service: String,
+    val options: List<Option>
+)