+++ pipecat-client-android/src/main/java/ai/pipecat/client/types/TransportState.kt
@@ -1,0 +1,15 @@
+package ai.pipecat.client.types
+
+/**
+ * The current state of the session transport.
+ */
+enum class TransportState {
+    Disconnected,
+    Initializing,
+    Initialized,
+    Authorizing,
+    Connecting,
+    Connected,
+    Ready,
+    Error
+}