+++ pipecat-client-android/src/main/java/ai/pipecat/client/utils/Json.kt
@@ -1,0 +1,12 @@
+package ai.pipecat.client.utils
+
+import ai.pipecat.client.types.Value
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.decodeFromJsonElement
+
+internal val JSON_INSTANCE = Json { ignoreUnknownKeys = true }
+
+internal fun <E> valueFrom(serializer: KSerializer<E>, value: E): Value {
+    return JSON_INSTANCE.decodeFromJsonElement(JSON_INSTANCE.encodeToJsonElement(serializer, value))
+}