+++ rtvi-client-android/src/main/java/ai/rtvi/client/helper/Legacy.kt
@@ -1,7 +1,0 @@
-package ai.rtvi.client.helper
-
-@Deprecated("VoiceClientHelper renamed to RTVIClientHelper")
-typealias VoiceClientHelper = RTVIClientHelper
-
-@Deprecated("RegisteredVoiceClient renamed to RegisteredRTVIClient")
-typealias RegisteredVoiceClient = RegisteredRTVIClient