+++ rtvi-client-android/src/main/java/ai/rtvi/client/types/MediaDeviceInfo.kt
@@ -1,9 +1,0 @@
-package ai.rtvi.client.types
-
-/**
- * Information about a media device.
- */
-data class MediaDeviceInfo(
-    val id: MediaDeviceId,
-    val name: String,
-)