+++ rtvi-client-android/src/main/java/ai/rtvi/client/types/MediaTrackId.kt
@@ -1,8 +1,0 @@
-package ai.rtvi.client.types
-
-/**
- * An identifier for a media track.
- */
-data class MediaTrackId(
-    val id: String
-)