+++ rtvi-client-android/src/main/java/ai/rtvi/client/types/ParticipantId.kt
@@ -1,8 +1,0 @@
-package ai.rtvi.client.types
-
-/**
- * A unique identifier for a session participant.
- */
-data class ParticipantId(
-    val id: String
-)