+++ rtvi-client-android/src/main/java/ai/rtvi/client/types/RTVIURLEndpoints.kt
@@ -1,6 +1,0 @@
-package ai.rtvi.client.types
-
-data class RTVIURLEndpoints(
-    val connect: String = "/connect",
-    val action: String = "/action"
-)