+++ rtvi-client-android/src/main/java/ai/rtvi/client/types/RegisteredHelper.kt
@@ -1,8 +1,0 @@
-package ai.rtvi.client.types
-
-import ai.rtvi.client.helper.RTVIClientHelper
-
-internal class RegisteredHelper(
-    val helper: RTVIClientHelper,
-    val supportedMessages: Set<String>
-)