+++ rtvi-client-android/src/main/java/ai/rtvi/client/types/ServiceConfigDescription.kt
@@ -1,9 +1,0 @@
-package ai.rtvi.client.types
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ServiceConfigDescription(
-    val name: String,
-    val options: List<OptionDescription>
-)