+++ rtvi-client-android/src/main/java/ai/rtvi/client/types/ServiceRegistration.kt
@@ -1,9 +1,0 @@
-package ai.rtvi.client.types
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ServiceRegistration(
-    val service: String,
-    val value: String
-)